# src/models/httpsuccess_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/swag"
)

// HttpsuccessResponse httpsuccess response
// swagger:model HTTPSuccessResponse
type HttpsuccessResponse struct {

	// data
	Data string `json:"data,omitempty"`
}

// Validate validates this httpsuccess response
func (m *HttpsuccessResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *HttpsuccessResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *HttpsuccessResponse) UnmarshalBinary(b []byte) error {
	var res HttpsuccessResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}